push handling into subcomponent
